moved redux ssr initializer to dashboard
